Sortting, Background, Fixing Roger complaints
Implemented various fixes to soothe the gripes and complaints of 1 (one) Roger Young. the terrorist known in some circles as the illusive "Dakamap".
When adding cards to a deck the user may specify quantity. This quantity is limited from 0-4 to enforce the standard ruleset.
Increased deck analytic percentages from 1% accuracy to 0.01% accuracy.
Added a message to the create card page that notifies the user of how lands will be treated.
Decks and cards within decks are now sorted alpabetically.
Background color changed to allow green text to be distinguishable